veritabanları fonksiyonlara atandı. front endde bazı veritabanı listelemeleri yapıldı.

+++ Admin/posts.php
@@ -2,7 +2,17 @@
 <?php include "Includes/_Sidebar.php"; ?>
 <div id="content-wrapper">
     <div class="container-fluid">
-      <a class="btn btn-large btn-primary my-4" style="color:white;" data-toggle="modal" data-target="#add_modal">Add Post</a>
+       <div class="row">
+         <a class="btn btn-large btn-primary p-3 offset-1 my-4 col-1" style="color:white;" data-toggle="modal" data-target="#add_modal">ADD POST</a>
+         <div class="alert alert-warning offset-1 col-2 my-4 alert-dismissible fade show" role="alert">
+           Total number of posts: <strong>
+
+             <?php $post_count=dbmyAdminPagePostsList();
+              $post_count=$con->query($post_count);
+              echo $post_count->num_rows;
+            ?></strong>
+         </div>
+       </div>
 
         <!-- DataTables -->
         <div class="card ">
@@ -35,39 +45,41 @@
             <tbody>
               <!--YAZILARI LİSTELEME-->
               <?php
-              $sql_list="SELECT * FROM posts";
-              $select_all_posts=mysqli_query($con,$sql_list);
+              $sql_list=dbmyAdminPagePostsList();
+              $sql_list=$con->query($sql_list);
               $k=1;
-              while ($row=mysqli_fetch_assoc($select_all_posts)) {
-                $post_id=$row["post_ID"];
-                $post_author=$row["post_AUTHOR"];
-                $post_author_role=$row["post_AUTHOR_ROLE"];
-                $post_date=$row["post_DATE"];
-                $post_title=$row["post_TITLE"];
-                $post_hide=$row["post_HIDE"];
-                $post_content=$row['post_CONTENT'];
-                $post_like_count=$row["post_LIKE_COUNT"];
-                echo "<tr>
-                    <td>{$post_id}</td>
-                    <td>{$post_title}</td>
-                    <td>{$post_date}</td>
-                    <td>{$post_author}</td>
-                    <td>".substr($post_content,0,20)."...</td>
-                    <td  style='width:5%'>
-                        <div class='dropdown'>
-                            <button class='btn btn-primary dropdown-toggle' type='button' id='dropdownMenuButton' data-toggle='dropdown' aria-haspopup='true' aria-expanded='false'>
-                                Actions
-                            </button>
-                            <div class='dropdown-menu' aria-labelledby='dropdownMenuButton'>
-                                <a class='dropdown-item' data-toggle='modal' data-target='#edit_modal$k' href='#'>Edit</a>
-                                <div class='dropdown-divider'></div>
-                                <a class='dropdown-item' href='posts.php?delete={$post_id}'>Delete</a>
+              if($sql_list->num_rows>0)
+              {
+                while ($row=$sql_list->fetch_assoc()) {
+                  $post_id=$row["post_ID"];
+                  $post_author=$row["post_AUTHOR"];
+                  $post_author_role=$row["post_AUTHOR_ROLE"];
+                  $post_date=$row["post_DATE"];
+                  $post_title=$row["post_TITLE"];
+                  $post_hide=$row["post_HIDE"];
+                  $post_content=$row['post_CONTENT'];
+                  $post_like_count=$row["post_LIKE_COUNT"];
+                  echo "<tr>
+                      <td>{$post_id}</td>
+                      <td>{$post_title}</td>
+                      <td>{$post_date}</td>
+                      <td>{$post_author}</td>
+                      <td>".substr($post_content,0,20)."...</td>
+                      <td  style='width:5%'>
+                          <div class='dropdown'>
+                              <button class='btn btn-primary dropdown-toggle' type='button' id='dropdownMenuButton' data-toggle='dropdown' aria-haspopup='true' aria-expanded='false'>
+                                  Actions
+                              </button>
+                              <div class='dropdown-menu' aria-labelledby='dropdownMenuButton'>
+                                  <a class='dropdown-item' data-toggle='modal' data-target='#edit_modal$k' href='#'>Edit</a>
+                                  <div class='dropdown-divider'></div>
+                                  <a class='dropdown-item' href='posts.php?delete={$post_id}'>Delete</a>
 
-                            </div>
-                        </div>
-                    </td>
-                </tr>";
-                ?>
+                              </div>
+                          </div>
+                      </td>
+                  </tr>";
+                  ?>
                 <!--YAZILARI LİSTELEME FINISH-->
 
                 <div id="edit_modal<?php echo $k; ?>" class="modal fade">
@@ -106,11 +118,11 @@
                                     </div>
                                     <div class="form-group">
                                         <label for="post_content">Content</label>
-                                        <textarea type="textarea" class="form-control" rows="10" required name="post_content"><?php echo $post_content ?></textarea>
+                                        <textarea type="textarea" class="form-control" rows="10" required name="post_content"><?php echo $post_content; ?></textarea>
                                     </div>
                                     <div class="form-group">
                                         <input type="hidden" name="post_author_role" value="<?php echo $post_author_role; ?>">
-                                        <input type="hidden" name="post_id" value="<?php echo $post_id?>">
+                                        <input type="hidden" name="post_id" value="<?php echo $post_id; ?>">
                                         <input type="submit" class="btn btn-primary" name="edit_post" value="Edit">
                                     </div>
                                 </form>
@@ -120,7 +132,7 @@
                     </div>
                 </div>
 
-                <?php $k++; } ?>
+                <?php $k++; }} ?>
 
             </tbody>
         </table>
@@ -171,9 +183,9 @@
         $post_author_role=$_POST["admin_role"];
         $post_date=date("d.m.Y")." ".date("H:i:s");
         $post_content=$_POST["post_content"];
-        $sql_add="INSERT INTO posts (post_AUTHOR,post_AUTHOR_ROLE,post_DATE,post_TITLE,post_CONTENT,post_LIKE_COUNT)
-        VALUES ('{$post_author}','{$post_author_role}','{$post_date}','{$post_title}','{$post_content}',0)";
-        $create_post_query=mysqli_query($con,$sql_add);
+        $sql_add=dbmyAdminPagePostsAdd($post_author,$post_author_role,$post_date,$post_title,$post_content);
+        $con->query($sql_add);
+        $con->close();
         header("Location: posts.php");
       }
        ?>
@@ -181,23 +193,27 @@
 
        <!--YAZI DÜZENLEME-->
          <?php
-            if(isset($_POST['edit_post']) && isset($_SESSION['role']) && $_SESSION['role']=='admin'){
+            if((isset($_POST['edit_post'])) && (isset($_SESSION['role'])) && ($_SESSION['role']=='admin')){
+              echo "<script>alert('echo $_SESSION[role]');</script>";
+              echo "<script>alert('girdima');</script>";
               $post_title=$_POST["post_title"];
               $post_content=$_POST["post_content"];
               $post_hide=$_POST["post_hide"];
               $sql_update="UPDATE posts SET post_TITLE='$post_title',
-              post_CONTENT='$post_content',post_HIDE=$post_hide WHERE post_ID='$_POST[post_id]'";
+              post_CONTENT='$post_content',post_HIDE='$post_hide' WHERE post_ID='$_POST[post_id]'";
               $edit_post_query=mysqli_query($con,$sql_update);
               header("Location:posts.php");
             }
             if(isset($_POST['edit_post']) && isset($_SESSION['role']) && $_SESSION['role']!='admin'){
+              echo "<script>alert('echo $_SESSION[role]');</script>";
               if($_POST['post_author_role']!='admin'){
+                echo "<script>alert('girdimb');</script>";
                 $post_title=$_POST["post_title"];
                 $post_content=$_POST["post_content"];
                 $post_hide=$_POST["post_hide"];
-                $sql_update="UPDATE posts SET post_TITLE='$post_title',
-                post_CONTENT='$post_content',post_HIDE='$post_hide' WHERE post_ID='$_POST[post_id]'";
-                $edit_post_query=mysqli_query($con,$sql_update);
+                $sql_update=dbmyAdminPagePostsEdit($post_title,$post_content,$post_hide,$_POST['post_id']);
+                $con->query($sql_update);
+                $con->close();
                 header("Location:posts.php");
               }
             }
@@ -208,8 +224,9 @@
        <?php
        if(isset($_GET["delete"]) && isset($_SESSION['role']) && $_SESSION['role']=='admin' ){
          $del_posts_id=$_GET['delete'];
-         $sql_delete="DELETE FROM posts WHERE post_ID={$del_posts_id}";
-         $delete_post_query=mysqli_query($con,$sql_delete);
+         $sql_delete=dbmyAdminPagePostsDelete($del_posts_id);
+         $con->query($sql_delete);
+         $con->close();
          header("Location:posts.php");
        }
         ?>
